requirements.txt
fastapi
uvicorn
ultralytics
Pillow
gunicorn
python-multipart
numpy
opencv-python
streamlit